requirements.txt
pytest
pytest-cov
colorama>=0.4.4
tqdm>=4.61.0
Pillow>=8.0.0